Added B-LDPC Codes implementation
B-LDPC Codes implementation in IT++ consists of a BLDPC_Parity and
BLDPC_Generator classes. The former one can be used for constructing
the parity-check matrix from so-called base-matrix.  Base matrix can
be saved and loaded from a text file. The BLDPC_Generator class
implements an efficient encoding algorithm using a preprocessed
version of the parity-check matrix.


git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/trunk@1062 a6533b66-300f-0410-b061-b36de6c4fb2d