Use copy_vector() where possible
This patch improves the consistency and possibly the performance
of various Vec's methods by using copy_vector() instead of
explicit copying of elements in a loop. It also fixes a minor bug
in "/=" operator for the case of division of a vector with itself,
i.e. "v /= v;". Finally, this patch adds missing friend
declarations for concat() functions and improves if statements in
the alloc() function.

git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/trunk@1091 a6533b66-300f-0410-b061-b36de6c4fb2d