Fixed regressions related to the specialised operator*()
The templated friend operator*() for two vectors, which calculates
the dot product, needs to be instantiated in the same way as the
dot() function. Besides, the declaration of the friend operator*()
included the definition. This definition is now moved below dot()
function definition. Without these two changes, a segmentation
fault occurred in vec_test under Cygwin.



git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/branches/itpp-4-0@1141 a6533b66-300f-0410-b061-b36de6c4fb2d