Improved memory allocation routines
Following the idea proposed in feature request [1277721]
implementation of create_elements() functions was improved. First
the memory for an Array, Vec or Mat is allocated and then the
constructor of each element is executed, but only if necessary.
For simple types like bin, int, double, etc., the constructor is
not invoked. Moreover, in the previous implementation the
constructor of each Array<Array>, Array<Vec> or Array<Mat> element
was invoked twice. Now this is fixed. Thanks to George Jongren for
submitting this feature request.

git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/branches/itpp-4-0@1177 a6533b66-300f-0410-b061-b36de6c4fb2d