Changed the way version numbers are handled
This patch fixes the problem of removing the "itpp.spec" and
"itpp-config.1" files during "make distclean". These files should
not be removed, since they are distributed in packages. Since now,
version numbers in these two files and in "configure.ac" are
updated according to a VERSION file content, when "autogen.sh"
bootstrapping script is executed.

git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/branches/itpp-4-0@1246 a6533b66-300f-0410-b061-b36de6c4fb2d