Add "--with-zdotu" option to select zdotu_ calling convention
The "--with-zdotu=<method>" option can be used to change the
default calling convention of BLAS zdotu_ Fortran function. The
BLAS libraries built with g77 compiler do not return the complex
result of a function, but pass it via the first argument of the
function. The implementations compiled with gfortran and other
vendor compilers return complex result as a typical function. By
using a "void" method with the "--with-zdotu=<method>" option, the
former approach is used (g77 compatible). To use the zdotu_ call
that returns the result by value, a "complex" method should be
passed. Without setting any of the two methods explicitly, IT++
tries to guess the proper method based on the detected BLAS
implementation. Otherwise, a Fortran wrapper function zdotusub_ is
used, which requires a working Fortran compiler. Finally, if this
is not the case, the zdotu_ BLAS method is not used at all and a
relevant warning message is displayed during the configuration
step.


git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/trunk@1428 a6533b66-300f-0410-b061-b36de6c4fb2d