Removed non-portable "--with-zdotu=complex" method
It seems that there is no portable solution for calling from C++
the Fortran functions that returns a complex value. C99 introduces
"double _Complex" type, which is compatible with Fortran COMPLEX.
But this C99 type is not a valid C++ type. Therefore this patch
removes the "complex" method from "--with-zdotu" option. The valid
methods are now:
- "zdotusub" - use a zdotusub_ Fortran wrapper function
- "void" - compatible with BLAS libraries built with g77 compiler
- "none" - do not use zdotu_ function at all


git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/branches/itpp-4-0@1431 a6533b66-300f-0410-b061-b36de6c4fb2d