Clean-up redundant use of "const" keyword in the Mat class
There is no need for declaring the returned by value types with an
additional "const" keyword. Moreover, Num_T is assumed a simple or
built in type for the Mat class, i.e. bin, short int, int, double
and complex<double>. Therefore it might be more efficient to pass
arguments of this type by value, in which case the "const" keyword
is also redundant.


git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/trunk@1480 a6533b66-300f-0410-b061-b36de6c4fb2d