Fix incorrect implementation of rint() function used under MSVC++
The rint() function is not available under MSVC++ compiler. Thus,
the local implementation has to be conditionally enabled for this
compiler. This patch fixes the incorrect implementation of rint()
function, which behaved like a similar round() function. It also
moves the "double rint(double)" definition out of the itpp
namespace.


git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/trunk@1520 a6533b66-300f-0410-b061-b36de6c4fb2d