Fix regression of a missing inline keyword in outer_product() specialisation
The inline keyword is necessary for template specialisations that
are defined in the header file. This bug resulted in linking
problems without any external BLAS library, when this particular
specialisation was used.

Thanks to Markus Dittrich for investigating this problem and
providing a patch for it on Gentoo Bugzilla (bug #217621).


git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/branches/itpp-4-0@1551 a6533b66-300f-0410-b061-b36de6c4fb2d