Fix bug [2004345] causing incorrect result of Mat::concat_vertical()
In case when one of the input matrices to concat_horizontal() or
concat_vertical() functions was empty, an error occurred. With
this patch the result of concatenating an empty matrix with
some other matrix is defined and a copy of the other matrix is
returned. This fixes bug [2004345].

Thanks to Carlos Pineda for reporting this problem.

git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/trunk@1589 a6533b66-300f-0410-b061-b36de6c4fb2d