String parser reimplemented to support multiple "a:b:c" format strings
The new implementation supports multiple "a:b[:c]" formats in one
string. For instance:
  ivec v = "1:2:5, 6 8 9:-1:0";
now works.

Memory handling is also more efficient, i.e. for the cases when
"a:b:c" format is not used, memory is allocated only once.
Besides, the new implementation should be easier to maintain,
because the code is much more modular.

This patch resolves feature request [2041480].

git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/trunk@1611 a6533b66-300f-0410-b061-b36de6c4fb2d